+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added YouTube Data API v3 fallback for notification titles: notification titles now use the Data API as a fallback.
+
 ## [2.5.0] - 2025-07-01
 
 ### Added